try this

+++ sw/oscillator.h
@@ -6,8 +6,8 @@ struct oscillator {
     uint16_t control;
     uint16_t wavetable;
     uint16_t resolution;
-    uint32_t audio_step;   /* software phase step: f * 2^32 / SAMPLE_RATE */
-    uint32_t phase;        /* software phase accumulator */
+    double   audio_step;   /* samples advanced per output tick (1.0 = native) */
+    double   phase;        /* current sample position into the wavetable */
     uint8_t note;
     bool    in_use;
 };
